renamed chart to figure
