+++ webviz_subsurface/plugins/_well_analysis/_figures/well_overview_figure.py
